Add PSN GitHub environments and secrets with Managed Identities (#603)
This pull request introduces significant enhancements to the
infrastructure provisioning for the production environment, primarily to
support a new "psn-prod" environment.
The changes add a new Azure subscription/provider configuration, manage
new user-assigned identities and federated credentials, and automate the
creation of GitHub Actions environment secrets for the "psn-prod"
environment.
Additionally, the repository module is configured to recognize the new
environment.

Resolves: CES-1567